backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
firecrawl-py
pydantic
pytest
pytest-asyncio
